Clear button & Random button changed
Clear button now focuses the input element when clicked, and does
nothing if the input element is empty. A stylistic choice so that it
seems to have one purpose and doesn't focus an empty input when clicked.

Random button now opens a random Wikipedia entry in a new tab when
clicked.

Perhaps later the Random button can open the random page inside of the
app.